+++ CHANGELOG.md
@@ -49,6 +49,8 @@ Improvements
 - Added an .active_plugin_header property to the WorkflowTree
 - The GenericTree.order_node_ids now also sets the active node again.
 - The WorkflowEditFrame now also allows to filter plugins for their name.
+- Added functions to fit circles and ellipses.
+
 
 Bugfixes
 --------